requirements.txt
# Bleeding edge Django
Django>=3.1,<3.2

# Django Debug
django-cors-headers
django-debug-toolbar

# Django-Extensions
django-extensions
Werkzeug
pyOpenSSL

# Web app to serve its own static files
whitenoise

# Configuration
django-environ

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth

# Time zones support
pytz

# Password storage
argon2-cffi

# RESTful
djangorestframework
django-rest-swagger
drf_yasg

# DB connection
# pre-requisite: sudo apt-get install python3-dev
mysqlclient

# HTML Content
Pillow
sorl-thumbnail
django-prettyjson
django-colorfield

# Django Admin Style
django-jazzmin

# My awesome package
django-onetimelink

# Squest
Markdown==3.3.4
PyYAML==5.4.1
fontawesome-free==5.15.2
celery==5.0.5
django-celery-results==2.0.1
towerlib==3.7.0
django-guardian==2.3.0
django-fsm==2.7.1
django-celery-beat==2.2.0
Pygments==2.8.1
django-filter==2.4.0
coverage==5.5
django-node-assets==0.9.8
mkdocs==1.1.2
mkdocs-material==7.1.2
pymdown-extensions==8.1.1
django-taggit==1.5.1
graphviz==0.16
# django-auth-ldap==2.4.0